changes to the launchfile making things easier to manipulate in the future
